POC/requirements.txt
flask
flask_cors
pytesseract
pdfplumber
python-docx
transformers
torch
sentence-transformers
faiss-cpu
